+++ app/ui/user/user-credit-result/user-credit-result.tsx
@@ -1,4 +1,4 @@
-import { fetchUser } from '@/app/business/services/auth';
+import { fetchUser } from '@/app/business/services/user/user.query';
 import books from '@/public/assets/books.png';
 import pencil from '@/public/assets/pencil.png';
 import Image from 'next/image';
